ambusync/backend/requirements.txt
# AmbuSync backend — Flask + MongoDB + Socket.IO + ML + PDF
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pymongo>=4.6.0
PyJWT>=2.8.0
bcrypt>=4.1.0
numpy>=1.26.0
scikit-learn>=1.4.0
fpdf2>=2.7.0
werkzeug>=3.0.0
google-generativeai>=0.8.0
